# openapi/QAlbum.yaml
openapi: 3.0.3
info:
  title: QAlbum API
  description: 家庭私有云相册后端 API（家庭空间版）
  version: 1.0.0
  contact:
    name: QAlbum Team
servers:
  - url: https://album.example.com/api
    description: Production server via Cloudflare Tunnel
paths:
...
  /auth/login:
    post:
      summary: 微信小程序登录
      description: 用 wx.login 获取的 code 换取服务端 Token
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code]
              properties:
                code:
                  type: string
                  description: 微信登录凭证
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  token:
                    type: string
                  expires_in:
                    type: integer
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ─────────────────────────────── 家庭空间 ───────────────────────────────
...
  /spaces/{space_id}:
    get:
      summary: 获取家庭空间详情
      tags: [Space]
      security:
        - BearerAuth: []
      parameters:
        - name: space_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Space'

    put:
      summary: 修改家庭空间（重命名、改配额、改备份路径）
      description: 仅管理员可操作
      tags: [Space]
      security:
        - BearerAuth: []
      parameters:
        - name: space_id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                quota_bytes:
                  type: integer
                backup_path:
                  type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Space'

    delete:
      summary: 删除家庭空间
      description: 仅管理员可删除，物理文件全部删除
      tags: [Space]
      security:
        - BearerAuth: []
      parameters:
        - name: space_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '204':
          description: 删除成功
...
  /spaces/{space_id}/albums:
    get:
      summary: 获取相册列表
      tags: [Album]
      security:
        - BearerAuth: []
      parameters:
        - name: space_id
          in: path
          required: true
          schema: { type: integer }
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: page_size
          in: query
          schema: { type: integer, default: 20 }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  page:
                    type: integer
                  page_size:
                    type: integer
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Album'

    post:
      summary: 创建相册
      description: 空间内任意成员均可创建
      tags: [Album]
      security:
        - BearerAuth: []
      parameters:
        - name: space_id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  maxLength: 50
      responses:
        '201':
          description: 创建成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Album'
  /spaces/{space_id}/albums/{album_id}:
    get:
      summary: 获取相册详情
      tags: [Album]
      security:
        - BearerAuth: []
      parameters:
        - name: space_id
          in: path
          required: true
          schema: { type: integer }
        - name: album_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Album'

    put:
      summary: 修改相册（重命名）
      description: 空间内任意成员均可操作
      tags: [Album]
      security:
        - BearerAuth: []
      parameters:
        - name: space_id
          in: path
          required: true
          schema: { type: integer }
        - name: album_id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Album'

    delete:
      summary: 删除相册
      description: 空间内任意成员均可删除，物理文件不受影响
      tags: [Album]
      security:
        - BearerAuth: []
      parameters:
        - name: space_id
          in: path
          required: true
          schema: { type: integer }
        - name: album_id
          in: path
          required: true
          schema: { type: integer }
      responses:
        '204':
          description: 删除成功

  # ─────────────────────────────── 文件 ───────────────────────────────
...
  /spaces/{space_id}/duplicates:
    get:
      summary: 检测重复图片
      description: 在空间维度跨相册同步检测，返回相似图片组
      tags: [Duplicate]
      security:
        - BearerAuth: []
      parameters:
        - name: space_id
          in: path
          required: true
          schema: { type: integer }
        - name: threshold
          in: query
          schema: { type: integer, default: 10 }
          description: 汉明距离阈值
      responses:
        '200':
          description: 检测完成
          content:
            application/json:
              schema:
                type: object
                properties:
                  groups:
                    type: array
                    items:
                      type: object
                      properties:
                        files:
                          type: array
                          items:
                            $ref: '#/components/schemas/File'
                        similarity:
                          type: integer
                          description: 组内最大汉明距离

  # ─────────────────────────────── 回收站 ───────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Space:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        owner_id:
          type: integer
          description: 创建者/管理员 ID
        quota_bytes:
          type: integer
        backup_path:
          type: string
          nullable: true
        backup_status:
          type: string
          enum: [ok, pending, unavailable]
        pending_backup_count:
          type: integer
          description: 待同步文件数量
        member_count:
          type: integer
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Album:
      type: object
      properties:
        id:
          type: integer
        space_id:
          type: integer
        name:
          type: string
        deleted_at:
          type: string
          format: date-time
          nullable: true
          description: 软删除标记，null 表示正常
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    File:
      type: object
      properties:
        id:
          type: integer
        space_id:
          type: integer
        album_id:
          type: integer
        filename:
          type: string
        file_type:
          type: string
          enum: [image, video]
        file_size:
          type: integer
        mime_type:
          type: string
        storage_path:
          type: string
          description: 相对 /data/photos/<space_id>/ 的路径
        thumbnail_path:
          type: string
          nullable: true
        cover_path:
          type: string
          nullable: true
          description: 视频封面路径
        phash:
          type: string
          nullable: true
          description: 64bit pHash 的 hex 字符串
        width:
          type: integer
          nullable: true
        height:
          type: integer
          nullable: true
        duration:
          type: integer
          nullable: true
          description: 视频时长（秒）
        uploaded_by:
          type: integer
        uploaded_at:
          type: string
          format: date-time
        deleted_at:
          type: string
          format: date-time
          nullable: true
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: string
  responses:
    BadRequest:
      description: 请求参数错误
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: 未认证或 Token 过期
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'